requirements.txt
requests==2.28.*
python-dotenv==1.0.*
python-telegram-bot==13.15
google-cloud-dialogflow==2.23.*
